[ISV-4964] Rename invalidation script
