+++ operator-pipeline-images/operatorcert/entrypoints/invalidate_preflight_versions.py
