+++ pyproject.toml
@@ -87,7 +87,7 @@ upload-artifacts = "operatorcert.entrypoints.upload_artifacts:main"
 upload-signature = "operatorcert.entrypoints.upload_signature:main"
 validate-catalog-format = "operatorcert.entrypoints.validate_catalog_format:main"
 verify-changed-dirs = "operatorcert.entrypoints.verify_changed_dirs:main"
-update-preflight-versions = "operatorcert.entrypoints.update_preflight_versions:main"
+invalidate-preflight-versions = "operatorcert.entrypoints.invalidate_preflight_versions:main"
 
 
 [build-system]
